versioncontrol: Merge pootle_app.vcs with versioncontrol.utils
