+++ pootle/apps/versioncontrol/utils.py
@@ -19,15 +19,13 @@
 
 """Utility functions to help with version control systems."""
 
-import os.path
+import os
 import shutil
 
-from translate.storage import versioncontrol
-
 from django.conf import settings
-
 from pootle_app.project_tree import to_podir_path
-from pootle_store.util import relative_real_path
+from pootle_store.util import add_trailing_slash, relative_real_path
+from translate.storage import versioncontrol
 
 
 class VersionControlError(Exception):
@@ -94,3 +92,51 @@ def add_files(path, files, message, author=None):
         shutil.copy(podir_path, vcs_path)
     output = vcs.add([to_vcs_path(f) for f in files], message, author)
     return output
+
+
+def recursive_files_and_dirs(ignored_files, ext, real_dir, file_filter):
+    """Traverses :param:`real_dir` searching for files and directories.
+
+    :param ignored_files: List of files that will be ignored.
+    :param ext: Only files ending with this extension will be considered.
+    :param real_dir:
+    :param file_filter: Filtering function applied to the list of files found.
+    :return: A tuple of lists of files and directories found when traversing the
+        given path and after applying the given restrictions.
+    """
+    real_dir = add_trailing_slash(real_dir)
+    files = []
+    dirs = []
+
+    for _path, _dirs, _files in os.walk(real_dir, followlinks=True):
+        # Make it relative:
+        _path = _path[len(real_dir):]
+        files += [os.path.join(_path, f) for f in filter(file_filter, _files)
+                  if f.endswith(ext) and f not in ignored_files]
+
+        # Edit _dirs in place to avoid further recursion into hidden directories
+        for d in _dirs:
+            if is_hidden_file(d):
+                _dirs.remove(d)
+
+        dirs += _dirs
+
+    return files, dirs
+
+
+def sync_from_vcs(ignored_files, ext, relative_dir,
+                  file_filter=lambda _x: True):
+    """Recursively synchronise the PO directory from the VCS directory.
+
+    This brings over files from VCS, and removes files in PO directory that
+    were removed in VCS.
+    """
+    if not hasversioning(relative_dir):
+        return
+
+    podir_path = to_podir_path(relative_dir)
+    vcs_path = to_vcs_path(relative_dir)
+    vcs_files, vcs_dirs = recursive_files_and_dirs(ignored_files, ext,
+                                                   vcs_path, file_filter)
+    files, dirs = recursive_files_and_dirs(ignored_files, ext, podir_path,
+                                           file_filter)
